+++ core/dbt/auth/__init__.py
@@ -1,14 +1,5 @@
 from dbt.auth.chain import OAUTH_CLIENT_ID, AuthChain  # noqa: F401
 from dbt.auth.credentials import Credential, CredentialKind, OAuthSession  # noqa: F401
-from dbt.auth.errors import (  # noqa: F401
-    AuthAborted,
-    AuthenticationExpired,
-    AuthError,
-    InadequateScopes,
-    InteractiveAuthError,
-    NotAuthenticated,
-    RefreshFailed,
-)
 from dbt.auth.resolvers import (  # noqa: F401
     CloudYamlResolver,
     EnvVarResolver,
@@ -17,3 +8,12 @@ from dbt.auth.resolvers import (  # noqa: F401
     ResolverKind,
 )
 from dbt.auth.session_cache import OAuthSessionCache  # noqa: F401
+from dbt.exceptions import (  # noqa: F401
+    AuthAborted,
+    AuthenticationExpired,
+    AuthError,
+    InadequateScopes,
+    InteractiveAuthError,
+    NotAuthenticated,
+    RefreshFailed,
+)
